implement rubah module

+++ build/module-utils/reconfigure.js
@@ -1,4 +1,23 @@
 "use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
+    Object.defineProperty(o, "default", { enumerable: true, value: v });
+}) : function(o, v) {
+    o["default"] = v;
+});
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -15,7 +34,10 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const fs_1 = __importDefault(require("fs"));
 const path_1 = __importDefault(require("path"));
 const readline_1 = __importDefault(require("readline"));
-const tsconfig_1 = require("./tsconfig");
+const tsconfig_1 = require("./templates/tsconfig");
+const child_process_1 = __importDefault(require("child_process"));
+const strings = __importStar(require("./templates/strings"));
+const jobs = __importStar(require("./templates/rubahjobs"));
 const crl = () => readline_1.default.createInterface({
     input: process.stdin,
     output: process.stdout,
@@ -24,25 +46,11 @@ const cwd = process.env.INIT_CWD;
 function reconfigure() {
     return __awaiter(this, void 0, void 0, function* () {
         console.log("reconfiguring tsemplate");
-        //SETTING UP RECONFIGURE SCRIPT
         let packageJson = JSON.parse(fs_1.default.readFileSync(path_1.default.join(cwd, "package.json")).toString());
-        packageJson.scripts.reconfigure = Object.keys(packageJson.averModule)
-            .sort((a, b) => packageJson.averModule[a].priority - packageJson.averModule[a].priority)
-            .filter(x => packageJson.averModule[x].reconfigure)
-            .filter(x => x != packageJson.name)
-            .map(x => packageJson.averModule[x].reconfigure)
-            .concat(["echo finished reconfiguring"])
-            .join(" && ");
         if (packageJson.averModule.tsemplate.firstrun) {
             delete packageJson.averModule.tsemplate.firstrun;
-            console.log(`
-tsemplate module has added few dependencies to your package.json
-please run 
-
-\u001b[42m\u001b[30mnpm run reconfigure && npm install\u001b[0m
-
-to install these dependencies and configure your project before continuing
-`);
+            console.log(strings.afterFirstReconfigure);
+            fs_1.default.writeFileSync(path_1.default.join(cwd, "README.md"), strings.readme1);
             fs_1.default.writeFileSync(path_1.default.join(cwd, "package.json"), JSON.stringify(packageJson, null, 2));
             return;
         }
@@ -63,6 +71,8 @@ to install these dependencies and configure your project before continuing
         if (mode == "module") {
             if (!fs_1.default.existsSync(path_1.default.join(cwd, "tsconfig.build.json")))
                 fs_1.default.writeFileSync(path_1.default.join(cwd, "tsconfig.build.json"), JSON.stringify(tsconfig_1.buildconfig, null, 2));
+            fs_1.default.writeFileSync(path_1.default.join(cwd, "README.md"), strings.readmeModule);
+            packageJson.scripts.reconfigure = "node -r ts-node/register src/module-utils/reconfigure.ts";
             packageJson.scripts['build:ts.d'] = "tsc -d --project tsconfig.build.json --emitDeclarationOnly";
             packageJson.scripts['build:ts'] = "tsc --project tsconfig.build.json";
             packageJson.scripts.build = "rm -rf build && mkdir build && npm run build:ts && npm run build:ts.d";
@@ -70,20 +80,36 @@ to install these dependencies and configure your project before continuing
             packageJson.main = "build/index.js";
             packageJson.bin = "build/index.js";
             packageJson.types = "build/index.d.js";
-            let index = fs_1.default.readFileSync(path_1.default.join(cwd, "src", "index.ts")).toString();
-            let i = index.indexOf("//your command line parsing logic here") + 38;
-            index = index.substr(0, i) + "\n        case 'reconfigure': require('./module-utils/reconfigure'); break;"
-                + index.substr(i);
-            fs_1.default.writeFileSync(path_1.default.join(cwd, "src", "index.ts"), index);
             if (!fs_1.default.existsSync(path_1.default.join(cwd, "src", "module-utils")))
                 fs_1.default.mkdirSync(path_1.default.join(cwd, "src", "module-utils"));
-            fs_1.default.writeFileSync(path_1.default.join(cwd, "src", "module-utils", "postinstall.ts"), "//YOUR POSTINSTALL SCRIPT HERE");
-            fs_1.default.writeFileSync(path_1.default.join(cwd, "src", "module-utils", "reconfigure.ts"), "//YOUR RECONFIGURE SCRIPT HERE");
+            if (!fs_1.default.existsSync(path_1.default.join(cwd, "src", "module-utils", "postinstall.ts")))
+                fs_1.default.writeFileSync(path_1.default.join(cwd, "src", "module-utils", "postinstall.ts"), "//YOUR POSTINSTALL SCRIPT HERE");
+            if (!fs_1.default.existsSync(path_1.default.join(cwd, "src", "module-utils", "reconfigure.ts")))
+                fs_1.default.writeFileSync(path_1.default.join(cwd, "src", "module-utils", "reconfigure.ts"), "//YOUR RECONFIGURE SCRIPT HERE");
             packageJson.scripts.postinstall = "node -r ts-node/register src/module-utils/postinstall.ts && " +
                 "node -r ts-node/register src/module-utils/reconfigure.ts";
         }
         else {
             // APPLICATION RECONFIGURATION GOES HERE
+        }
+        //initializing rubah folder
+        if (!fs_1.default.existsSync(path_1.default.join(cwd, ".avermodule")))
+            fs_1.default.mkdirSync(path_1.default.join(cwd, ".avermodule"));
+        if (!fs_1.default.existsSync(path_1.default.join(cwd, ".avermodule", "rubah")))
+            fs_1.default.mkdirSync(path_1.default.join(cwd, ".avermodule", "rubah"));
+        let jobsPath = path_1.default.join(cwd, ".avermodule", "rubah", "jobs");
+        if (!fs_1.default.existsSync(jobsPath))
+            fs_1.default.mkdirSync(jobsPath);
+        //Setting Up basic rubah jobs
+        fs_1.default.writeFileSync(path_1.default.join(jobsPath, "package.json"), JSON.stringify(jobs.packagejson));
+        fs_1.default.writeFileSync(path_1.default.join(jobsPath, "index.json"), JSON.stringify(jobs.index));
+        fs_1.default.writeFileSync(path_1.default.join(jobsPath, "readme.json"), JSON.stringify(jobs.readme));
+        for (let moduleName in packageJson.averModule) {
+            if (moduleName == packageJson.name)
+                continue;
+            let pkgjson = JSON.parse(fs_1.default.readFileSync(`node_modules/${moduleName}/package.json`).toString());
+            console.log(`reconfiguring ${moduleName}`);
+            console.log(child_process_1.default.execSync(`node_modules/${moduleName}/${pkgjson.main} reconfigure`).toString());
         }
         fs_1.default.writeFileSync(path_1.default.join(cwd, "package.json"), JSON.stringify(packageJson, null, 2));
     });
